+++ src/devices/innr.ts
@@ -2,9 +2,20 @@ import * as fz from "../converters/fromZigbee";
 import * as exposes from "../lib/exposes";
 import * as m from "../lib/modernExtend";
 import * as reporting from "../lib/reporting";
-import type {DefinitionWithExtend} from "../lib/types";
+import type {DefinitionWithExtend, Fz, KeyValueAny} from "../lib/types";
 
 const e = exposes.presets;
+
+const fzLocal = {
+    rc_110_level_to_scene: {
+        cluster: "genLevelCtrl",
+        type: ["commandMoveToLevel", "commandMoveToLevelWithOnOff"],
+        convert: (model, msg, publish, options, meta) => {
+            const scenes: KeyValueAny = {2: "1", 52: "2", 102: "3", 153: "4", 194: "5", 254: "6"};
+            return {action: `scene_${scenes[msg.data.level]}`};
+        },
+    } satisfies Fz.Converter<"genLevelCtrl", undefined, ["commandMoveToLevel", "commandMoveToLevelWithOnOff"]>,
+};
 
 export const definitions: DefinitionWithExtend[] = [
     {
@@ -958,7 +969,7 @@ export const definitions: DefinitionWithExtend[] = [
         model: "RC 110",
         vendor: "Innr",
         description: "Innr RC 110 Remote Control",
-        fromZigbee: [fz.command_step, fz.command_move, fz.command_stop, fz.command_on, fz.command_off, fz.rc_110_level_to_scene],
+        fromZigbee: [fz.command_step, fz.command_move, fz.command_stop, fz.command_on, fz.command_off, fzLocal.rc_110_level_to_scene],
         toZigbee: [],
         meta: {multiEndpoint: true},
         endpoint: (device) => {
